+++ src/parser/CDTParser.h
@@ -48,7 +48,7 @@ public:
   ~CDTParser(){};
 
   //
-  int parse(const char *buffer, unsigned int size);
+  int64_t parse(const char *buffer, unsigned int size);
 
   // To be iterated over in processing thread
   std::vector<struct CDTData> Result;
