+++ src/parser/R5560Parser.h
@@ -56,7 +56,7 @@ public:
   ~R5560Parser(){};
 
   //
-  int parse(const char *buffer, unsigned int size);
+  int64_t parse(const char *buffer, unsigned int size);
 
   // To be iterated over in processing thread
   std::vector<struct R5560Data> Result;
